+++ packages/core/src/models/curriculum/index.ts
@@ -2,9 +2,9 @@ import { schema } from '@nawadi/db'
 import { SQL, and, desc, eq, isNotNull } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useQuery } from '../../contexts/index.js'
-import { zod } from '../../util/zod.js'
+import { useQuery } from '../contexts/index.js'
 import { Module } from '../program/module.js'
+import { zod } from '../util/zod.js'
 
 export { Competency } from './competency.js'
 export * as Curriculum from './index.js'
