+++ packages/core/src/models/discipline.test.ts
@@ -1,17 +1,16 @@
 import assert from 'assert'
 import test from 'node:test'
-import { withTestTransaction } from '../../contexts/index.js'
-import { createDegree, listDegrees } from './degree.js'
+import { withTestTransaction } from '../contexts/index.js'
+import * as Discipline from './discipline.js'
 
-test('degree crud', () =>
+test('discipline crud', () =>
   withTestTransaction(async () => {
-    const { id } = await createDegree({
+    const { id } = await Discipline.create({
       title: 'title-1',
       handle: 'handle-1',
-      rang: 1,
     })
 
-    const list = await listDegrees()
+    const list = await Discipline.list()
 
     assert.equal(list.length, 1)
     const [item] = list
@@ -20,6 +19,5 @@ test('degree crud', () =>
       id,
       title: 'title-1',
       handle: 'handle-1',
-      rang: 1,
     })
   }))
