+++ packages/core/src/models/discipline.ts
@@ -1,8 +1,8 @@
 import { schema } from '@nawadi/db'
-import { useQuery } from '../../contexts/index.js'
-import { normalizeHandle, normalizeTitle, singleRow } from '../../util/index.js'
+import { useQuery } from '../contexts/index.js'
+import { normalizeHandle, normalizeTitle, singleRow } from '../util/index.js'
 
-export async function listDisciplines() {
+export async function list() {
   const query = useQuery()
 
   const rows = await query
@@ -16,10 +16,7 @@ export async function listDisciplines() {
   return rows
 }
 
-export async function createDiscipline(item: {
-  title: string
-  handle: string
-}) {
+export async function create(item: { title: string; handle: string }) {
   const query = useQuery()
 
   const rows = await query
