+++ packages/core/src/models/index.ts
@@ -1,6 +1,7 @@
 export * as Actor from './actor.js'
 export * as Curriculum from './curriculum/index.js'
 export * as Location from './location/index.js'
-export * as Program from './program/index.js'
 
-export * from './programs2/index.js'
+export * as Degree from './degree.js'
+export * as Discipline from './discipline.js'
+export * as Program from './program.js'
