+++ packages/core/src/models/program.test.ts
@@ -1,31 +1,31 @@
 import assert from 'assert'
 import test from 'node:test'
-import { withTestTransaction } from '../../contexts/index.js'
-import { createDegree } from './degree.js'
-import { createDiscipline } from './discipline.js'
-import { createProgram, listPrograms } from './program.js'
+import { withTestTransaction } from '../contexts/index.js'
+import * as Degree from './degree.js'
+import * as Discipline from './discipline.js'
+import * as Program from './program.js'
 
 test('program crud', () =>
   withTestTransaction(async () => {
-    const { id: disciplineId } = await createDiscipline({
+    const { id: disciplineId } = await Discipline.create({
       title: 'discipline-1',
       handle: 'dc1',
     })
 
-    const { id: degreeId } = await createDegree({
+    const { id: degreeId } = await Degree.create({
       title: 'degree-1',
       handle: 'dg1',
       rang: 1,
     })
 
-    const { id } = await createProgram({
+    const { id } = await Program.create({
       title: 'program-1',
       handle: 'pr1',
       degreeId,
       disciplineId,
     })
 
-    const list = await listPrograms()
+    const list = await Program.list()
 
     assert.equal(list.length, 1)
     const [item] = list
