+++ packages/core/src/models/program.ts
@@ -1,9 +1,9 @@
 import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
-import { useQuery } from '../../contexts/index.js'
-import { normalizeHandle, normalizeTitle, singleRow } from '../../util/index.js'
+import { useQuery } from '../contexts/index.js'
+import { normalizeHandle, normalizeTitle, singleRow } from '../util/index.js'
 
-export async function listPrograms() {
+export async function list() {
   const query = useQuery()
 
   const rows = await query
@@ -28,7 +28,7 @@ export async function listPrograms() {
   return rows
 }
 
-export async function createProgram(item: {
+export async function create(item: {
   title: string
   handle: string
   disciplineId: string
