+++ packages/core/src/models/program/competency.test.ts
@@ -1,6 +1,6 @@
 import assert from 'assert'
 import test from 'node:test'
-import { withTestTransaction } from '../../contexts/index.js'
+import { withTestTransaction } from '../contexts/index.js'
 import { Competency } from './competency.js'
 
 test('create competency', () =>
