+++ packages/core/src/models/program/competency.ts
@@ -2,8 +2,8 @@ import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useQuery } from '../../contexts/index.js'
-import { zod } from '../../util/zod.js'
+import { useQuery } from '../contexts/index.js'
+import { zod } from '../util/zod.js'
 
 export * as Competency from './competency.js'
