--- packages/core/src/models/programs2/index.ts
@@ -1,3 +1,0 @@
-export * from './degree.js'
-export * from './discipline.js'
-export * from './program.js'
